pyproject.toml
[project]
name = "django-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.6",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "fastapi>=0.117.1",
    "python-jose[cryptography]>=3.5.0",
    "requests>=2.32.5",
    "uvicorn>=0.36.0",
]

[dependency-groups]
dev = [
    "ruff>=0.13.1",
]
